Create React app
Use Create React App[1] to set up the initial framework:

    npx create-react-app data-viewer --template typescript

After running that, I moved everything out of the resulting
subdirectory, and merged the relevant parts of .gitignore and README.md.
Everything else is the original, out-of-the box version.

Although `npm audit` reports a security vulnerability
(CVE-2021-3803[1]), it appears not to affect this configuration due to
that the way the nth-check dependency is used: we are not parsing
arbitrary, user-provided CSS, but rather trusted CSS that we author. See
also "npm audit: Broken by Design"[3], explaining the problems with
auditing front-end code.

[1] https://create-react-app.dev/docs/getting-started
[2] https://nvd.nist.gov/vuln/detail/CVE-2021-3803
[3] https://overreacted.io/npm-audit-broken-by-design/

Issue #1 Set up this repository with development environment