Split out dev dependencies
There is not a strong difference between dependencies and dev
dependencies in front-end code, as all are needed to build, and the
resulting build artifact has all the dependencies it cares about built
in. However, I still find it useful to keep libraries that are used by
the application separate from libraries that are used to develop the
application, just for clarity.

Issue #1 Set up this repository with development environment