Add and use Pino for logging
Pino is a logging library, with a variety of transports, and which works
in both Node.js and the browser. We're currently using it on the
back-end[1], so let's also use it on the front-end.

For now, we'll just log to the browser's console window; someday we
might wire up a service like Data Dog, New Relic, or Sentry. Setting up
the logging framework now will let us write logging code as we build out
functionality, and wire it up later.

Unless overridden with the `REACT_APP_LOG_LEVEL` environment variable,
only errors will be logged in production builds, and INFO and above will
be logged in dev builds.

As part of wiring this up, fix several linter errors:
- @typescript-eslint/non-nullable-type-assertion-style
- @typescript-eslint/comma-dangle
- @typescript-eslint/no-floating-promises
- object-curly-newline

[1] https://github.com/PhilanthropyDataCommons/service/blob/main/src/logger.ts

Issue #1 Set up this repository with development environment
Issue #2 Set up CI