Prefer arrow functions
Classic JavaScript functions have some confusing and surprising behavior
built in, including `this` binding and the magic `arguments` object.

As a policy, I prefer arrow functions to classic functions, and intend
to code in this repo accordingly. However, due to tooling limitations, I
will not broadly enforce that in ESLint. There is a specific rule for
React components, though, which this commit configures.

Configure the react/function-component-definition[1] to require arrow
functions for both named and unnamed React components, and rewrite the
<App> component to be an arrow function expression.

[1] https://github.com/jsx-eslint/eslint-plugin-react/blob/master/docs/rules/function-component-definition.md

Issue #12 Implement basic interface elements