Replace placeholder text
Replace the "learn React" placeholder text with the title of the
application, and links to the repo and the website. Again, this is
primarily for the benefit of observers who want to see more visible
progress; it's still placeholder text and will be replaced as we build
out more functionality.

Also, I like using the spinning React logo as a modern "under
construction" logo. 🚧

Tagging this as part of the continuous deployment issue because one of
my goals is to see this up on the internet, hopefully both with
Netlify's deploy previews and then on main once it's merged, in case
configuring DNS takes longer that it takes to review this.

Issue #9 Set up CD