Require default arguments, not default props
@jasonaowen and I have decided that our pattern for component definition
will be to use TypeScript interfaces for describing the props required
by a component and, rather than using `defaultProps`, require that
function definitions provide default arguments instead. This seems more
idiomatic with modern JavaScript and TypeScript when the component is
not intended to be used outside of the project (I.e., as part of a
library).

Issue #27 Set up Storybook for documenting front-end components