Install and configure Storybook
This commit follows the basic instructions for installing Storybook[1]
to manage our front-end components and makes a few changes to make it
compatible with our codebase.

We used the `storybook@next` installation to get the v7 beta, which
resolved an initial bunch of problems using the current (v6) version.
Everything included in this commit was auto-generated from their `init`,
with the following changes:

- Disabled telemetry[2]
- Modified the sample components/stories to remove lint violations
- Configured ESLint’s `import/no-extraneous-dependencies` rule to let
  stories look in `devDependencies` for its dependencies
- While Storybook’s `init` added an override block to disable the
  `import/no-anonymous-default-export` rule for stories, I manually
  added an `import/no-default-export` override, since Storybook relies
  on default exports to function.
- Use a more idiomatic `build:storybook` command.

[1] https://storybook.js.org/
[2] https://storybook.js.org/docs/react/configure/telemetry

Issue #27: Set up Storybook for documenting front-end components