Disallow all robots
The data-viewer requires a signed-in user for all content; there are
currently no planned public pages. There's nothing for a search engine
to index, because every page will just be a redirect to the IdP -
assuming the crawler is smart enough to run a React app!

Issue #10 Scaffold routing and page construction