Set all elements to use border-box box-sizing
Setting the box-sizing[1] to `border-box` allows for consistent
definition of element sizes using their outer visible dimensions (from
outside edges of the border) rather than the browser default content
dimensions (exclusive of padding or border).

This is a standard CSS override and we all wish browsers would default
to this, if it wouldn’t break the web.

[1] https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing

Issue #12 Create basic interface components