Log in through OpenID Connect
Use the `@axa-fr/react-oidc` library[1] to allow logging in via OpenID
Connect. Add a <User> component to the <Header> that will show either a
login button or the current user's name.

Note that the login button will probably not last long-term, but until
we start showing any real data, it's a useful placeholder. It also looks
good during demos, rather than the automatic login I expect we'll want.
The <User> component is also unstyled for the moment; this commit is
more focused on the functionality.

This adds a fair bit of required configuration. Add an `.env.example`
file documenting several environment variables, both new and existing,
and both optional and required.

The library supports service-worker-only communication between the app
and the API, which is intended to improve security; at least for the
moment, do not opt in, as we still need to figure out how to handle
authentication in PR deploy previews. This may be something we want
eventually, at which point we should commit the currently-ignored
service worker & configuration files.

I did work far enough forward to confirm that I could make authenticated
requests to the API, but am not including that yet both for the sake of
scope and because it depends on unmerged service-side changes.

[1] https://www.npmjs.com/package/@axa-fr/react-oidc

Issue #6 Proof of concept authentication via Keycloak