Build and publish Storybook in CI & CD
When we first introduced Storybook, we had considered whether or not to
publish the component previews it provides as part of our continuous
delivery system, and initially decided to defer that work. Now that
we're starting to use it in earnest, let's make it easier for others to
see the value it provides!

Update the `build` script, which is invoked by both our continuous
integration and continuous delivery systems, to build both the
application and the Storybook content. This should be published under
`/storybook`.

Issue #2 Set up CI
Issue #9 Set up CD
Issue #27 Set up Storybook for documenting front-end components
PR #28 Add and configure Storybook