Fix Button to use correct type for children
Previously, I misunderstood how to indicate that an element received
standard React nodes as children. Consequently, using the `Button`
component *required* its children be wrapped in a fragment if you wanted
to mix JSX elements and strings.

This problem was masked by how our stories wrapped children in a single
fragment.

This commit fixes that, uses the correct `ReactNode` type, and updates
the story to use a syntax that replicates reality (in a weird way).

Issue #12 Create basic interface components