Position app frame with inset instead of 100vw/vh
The vw/vh units are vulnerable to the appearance of a scrollbar; if you
make an element 100vw wide, and then the viewport spawns a vertical
scrollbar, your element will now be wider than the available space,
spawning a horizontal scrollbar. Agony.

Positioning it as absolute with an inset of 0 is a “safer” way of
accomplishing the intended effect.

Issue #12