Use fixed and accessible spacing variables
Some spacing doesn’t need to increase when the user has indicated an
accessibility need through a larger root font-size; some spacing does.

This commit adds a handful of CSS variables attuned to each need: those
prefixed with `fixed-spacing` don’t change as the user’s font size
changes, and are most useful when UI does not need additional room as
text size increases. Those prefixed with `accessible-spacing` do grow
larger with text size, such as when we need elements to be farther apart
to be visually distinguishable.

This commit also removes the `font-size` property from being applied to
the `body` element, because we want the user’s preferences to rule.

Issue #12