Add ProposalTable stub component
This commit demonstrates how to bring the Panel and Table components
together to construct a ProposalTable, including a story for rendering
it.

I’m calling this a stub component because it contains static values
rather than being composed by props, which is of course not correct.

Issue #19 Create proposal detail page