Add ProposalTable component
Bring the Panel and Table components together to construct a
ProposalTable, including a story for rendering it.

Currently, it only allows for two columns, and the shape of its props
match that. As we add more functionality, the data it expects will grow
in complexity.

This is not yet used anywhere, but will be used in a subsequent commit.

Issue #19 Create proposal detail page

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>