Add support for clickable table rows
Some table rows will need to be… clickable! E.g., clicking on them will
perform some action like navigating to an entity detail page.

When `onClick` is passed to a `TableRow`, we apply it (of course) and
add some styling to indicate that it is clickable.

Issue #14 Create proposals table page