Add ProposalListTable component and story
A `ProposalListTable` generates a table from an array of `proposals`.
Each proposal is an object with the field `shortCode` as the key and the
array of values that the proposal has for that field as the value. (Yes,
the PDC supports multiple values for each field.)

Two additional props:

1. `fieldNames` map shortCodes to labels, for generating the header row.
2. `columns` defines which fields we’re showing and dictates their
   order, as it’s just an array of `shortCodes`.

This component does not have any awareness of pagination, as that will
be the responsible of a parent component or page. It will, eventually,
need to understand field visibility (#18) and sort order (#16).

This commit also includes a Storybook story.

Issue #14 Create proposals table page

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>