Add ProposalListTablePanel component
This adds the component for the actual panel wrapped around a
`ProposalListTable`, although until we add more functionality it is…
quite sparse.

E.g., in this iteration the `PanelHeader` has nothing but a title, which
in the near future will be replaced with `PanelAction` components.

Issue #14 Create proposals table page

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>