Add Paginator component
This commit adds the `Paginator` component and its story, as well as the
`PaginatorWrapper` that we use when actually placing the paginator on a
real body. I will separately document the use of this.

Note that the `Paginator` component file defines several sub-components
that make the file rather unwieldy. We can break those out to their own
files if we’d like, ala `Tabel` and `Panel`.

There are also two utility methods that could easily be moved to a
utility file, but we don’t yet have a pattern for that.

Issue #14 Create proposals table page
Issue #15 Implement pagination on proposal lists