Add Paginator component
This commit adds the `Paginator` component and its story, as well as the
`PaginatorWrapper` that we use when actually placing the paginator on a
real body.

There are also two utility methods that could easily be moved to a
utility file, but we don’t yet have a pattern for that.

Issue #15 Implement pagination on proposal lists