Add proposal detail page
Add a new <ProposalDetailPage> component, and wire it up at
`/proposal/:id`. This component will fetch data from the API - whose
location must be configured via an environment variable - and display
it.

The source of the header data is an educated guess, and we may need to
iterate on which canonical fields we pull out.

Note that this does not have any error handling yet: if you navigate to
a path with an invalid proposal ID, and the API 404s, the loading state
will be stuck until you navigate away. There will be an error in the
console log, and given the pilot nature of this project, that should
suffice until we have a chance to clean this up.

[1] https://react.dev/learn/reusing-logic-with-custom-hooks

Issue #19 Create proposal detail page