Add <ProposalList> page
Add a new page that will retrieve the list of proposals from the API,
transform the data into what is expected by the <ProposalListPanel>
component, and render that component when the data is ready.

Currently, the API does not provide the necessary details in the
proposal list response, so we have to do an N+1 query where we get the
details for each proposal in the list. This can take a long time and be
a lot of traffic, but until the API offers more information, it's the
only thing to do.

This has some pagination support, via the `page` and `count` query
parameters, like `/proposals?count=25&page=3`. These parameters are
passed along to the API. However, because the API also does not return
any information about how many proposals or how many pages there are,
set the default `count` parameter to be very high. Again, until the API
offers more information, the other alternatives lead to a bad user
experience.

Issue #14 Create proposals table page