Simplify syntax for adding an icon to a PanelTag
Previously, PanelTags had no interface for indicating an icon. You just
added it as a child in the component. You still can, but in preparation
for adding little badges (and because it’s a slightly more ergonomic
interface), this commit lets you set the icon using an `icon` property.
This also automatically adds the necessary `className` so you don’t have
to do it yourself each time.