Add badges to PanelTags
Following on the previous introduction of badges, this commit lets you
add a badge to a PanelTag with the `badge` property, which prefixes the
tag with a badge instead of an icon.

Technically, nothing stops you from adding a badge *and* an icon, in
which case the icon would appear first, but this is an undesired design.
I’m just not disallowing it because, well, that would be more code and
*maybe* you’ll find a good reason?