Add ability to toggle wrapping on proposal list
This commit adds the button that toggles wrapping on the proposal list
page. A future commit will add this to the proposal detail page, but
it’s less important there.

The button relies on the concept of a "truncated" table, or, one where
the contents of body cells don’t wrap and are held to a max-width, above
which they have an ellipsis.

Issue #53 Implement wrap toggling