Show previous/next proposal buttons in dev mode
Navigating by proposal ID is not an interface we want to show to users:
there is no semantic meaning to the proposal ID, and it depends on the
unsafe and incorrect assumption that there are no gaps in the sequence.
Long-term, we plan to build a narrow list on the left side of the detail
page that will allow navigating from one proposal to another.

However, developers do not really care about the lack of semantic
meaning around proposal IDs; our concerns are more around how the data
is displayed than what the data means. For that, quickly navigating
through multiple proposals - in any arbitrary but consistent order - is
very convenient.

Add previous/next links to the proposal detail page when run in dev mode
(`npm run start`). Do not show these links when run in production mode
(`npm run build`).