Allow configuration of column actions alignment
This commit lightly refactors the ColumnHead to allow for configuring
the alignment to be on the right side of the cell (default) or next to
the content (“left”).

Even while building this I thought of some additional configuration and
semantics I’d like to add, but I’m punting on those until needed. Just
know the ergonomics of this may change.