Style navbar user item like other items
Previously, we used the `<Button>` component for all variants of the
user item in the navbar. That was semantically wrong for three of them
anyway, and it also made it difficult to match the styling of other
navbar items.

This commit:

- Composes the item manually instead of using `<Button>` so we have more
  styling control
- Only uses a button for the variant that is actually interactive
- Applies navbar item styling to all variants
- Limits interactive properties (cursor) on the navbar items to only
  those items that are actually interactive (links, buttons)

Issue #61 Style login/user item in navbar like other links