Change proposal detail URL to `/proposals/{id}`
Previously, we used `/proposals/` as the URL for the proposals index and
`/proposal/{id}` (note the singular) for the proposal detail page. This
is sensible and defensible.

However, using a consistent path that either displays the index (if no
ID is specified) or individual entity (if one is) has become idiomatic
for RESTful web apps, e.g. GitHub (`/issues/{id}`) and StackOverflow
(`/questions/{id}`). It’s also slightly more ergonomic: you just add or
remove an ID to switch between index and individual, with no dangling
404 page like you get when you remove the ID from `/proposal/{id}`.

Issue #63 Use /proposals/{id} instead of /proposal/{id} for proposal
          detail URL