Add OffsiteLink component
This component makes it easy to responsibly link offsite. It has the
same ergonomics as the React Router `<Link>` component — e.g. using the
`to` attribute to supply your destination — but it defaults to opening
in a new tab/window, and adds the `rel` attribute for linking offsite
without leaking user data.[1]

[1] https://developer.chrome.com/docs/lighthouse/best-practices/external-anchors-use-rel-noopener/