Go to proposals page from log in button
Change the target of the two login buttons - in the `<User>` component
of the navbar, and in the landing page - to go to `/proposals` instead
of back to the landing page.

This does not affect the existing behavior of navigating to a secured
page while signed out, which will continue to go through the login
process and then redirect to the destination secured page.

Issue #75 Go to Dashboard after initial login