Handle API change to use bundles
In PR service#296[1], the return format of the `/proposals` API endpoint
is changing from a top-level array of proposals to an object containing
proposals under the `entries` key. It also returns the `total` number of
proposals, which will allow us to do pagination, but that's separate
work; for now, just update the parsing code to use the new format.

[1] https://github.com/PhilanthropyDataCommons/service/pull/296