Improve sticky table headers
Our table heads have top-borders. When the header is sticky, that top
border used to remain visible when the header was sticky. That’s not
ideal; a border is no longer needed to distinguish it from other
content.

This commit makes the border defined via a CSS variable rather than
statically, and sets the `top` property of the sticky header to the
negative value of that variable so that the header scrolls that much out
of view.