Extract <Search> component
In order to make the <input> a controlled component[1], it needs some
state. In order to avoid rerendering the entire proposal table on each
keystroke, we want to move that state below the <ProposalListTablePanel>
component. Extract a new <Search> component, and wire it up with some
placeholders.

Using an HTML <form>, an <input type=search>, and a submit button allows
us to leverage some of the default browser behavior. Hitting the enter
key after typing a query should cause the form to submit and the
`onSearch` prop function to fire, rather than requiring users to use the
mouse to click the button.

The only <input> in the application was the search box, so now that it's
part of its own component, move its styling into the component's CSS.

Note that as part of this change, the "Search" button is enabled, even
though it doesn't do anything yet. That'll come in the next commit.

[1] https://react.dev/reference/react-dom/components/input#controlling-an-input-with-a-state-variable

Issue #83 Enable search