Search loaded proposals by simple string matching
In order to meet some demo deadlines, and while the API implementation
is being figured out, implement "search" on the front-end as simple
string matching. As we intend the back-end to do, this shows any
proposal containing a field value that matches the search query; unlike
what we intend the back-end to do, it does no stemming, is not aware of
word boundaries, includes stopwords, and so on - it's strictly string
matching, with some lowercasing thrown in for case insensitivity.

This is intended to be reverted once the API supports searching, both
because of the additional functionality it will offer and because this
approach precludes pagination.

Issue #83 Enable search