Remove contribution prompt from login landing page
It's better to keep the login landing page as simple and free of
distractions as possible, since some newcomers might actually need to
read it carefully to figure out what to do next.  We can put
contribution information on other PDC pages, though: for example on
informational pages about the PDC, and perhaps in the data viewer
itself just elsewhere than the login landing page.

Resolves #88