Use deep proposals data from API
Recently[1], the API gained the ability to include all the necessary
data in the `/proposals` endpoint, instead of only including IDs. As a
result, we can stop making an additional N queries to `/proposal/{id}`
to fill out the list, significantly improving load times.

We're still not using the also-recently-introduced[2] `total` field,
which was introduced to support pagination, because instead we're doing
client-side search[3]. Once server-side search is ready[4], we'll be
able to use that and implement pagination.

[1] https://github.com/PhilanthropyDataCommons/service/pull/313
[2] https://github.com/PhilanthropyDataCommons/service/pull/296
[3] https://github.com/PhilanthropyDataCommons/data-viewer/pull/84
[4] https://github.com/PhilanthropyDataCommons/service/pull/320