Return complete `/proposals` response
In commit 1eb622009fe5c695ac8d10cd9b89321ce5951425, we updated the
client to continue to work with the new bundle format the API now
returns. However, for expediency and simplicity, we did not expose any
of the new data (currently, the total number of proposals).

Additionally, in commit dddd5417ecb0ca1ac5534fca86e5fae4e18a223c, we
started taking advantage of the new deep proposals data the API now
returns. The additional per-proposal requests we needed to make was what
kept us from reusing the `usePdcApi` effect; now that it's only a single
request, we'll be able to lean on that instead.

When server-side search is available, we'll need to be able to
distinguish between "we haven't loaded data yet" and "we've loaded the
data, and there are no search results". We couldn't make that
distinction when `useProposals` returned an empty array while awaiting
the API request to complete.

As a first step towards both pagination and server-side search, return
the full response from the `/proposals` API.

Issue #15 Implement pagination on proposal lists
Issue #83 Enable search