Refactor `useProposals` to use `usePdcApi`
Now that the `useProposals` effect is returning the API response as-is
rather than modifying it, and not needing to make subsequent requests
since the API is returning full proposals data from `/proposals`, we
don't need to duplicate the code in `usePdcApi`. Refactor the
`useProposals` effect to be a call to `usePdcApi`.