Add utility for picking best proposal value
We occasionally need to pick a proposal value from multiple candidate
fields. E.g., the organization name may be stored in one of several
fields; we’d like to go through them in a particular preferential order
and select the first valid one.

This commit adds `getProposalValuesFromCandidates` for doing just that,
along with tests.

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>