Stop exporting internal `usePdcApi` function
Now that we have three uses of it, I think we can safely say that the
`usePdcApi` hook is only intended to be wrapped in a per-endpoint helper
method. There are no external uses of it - and I don't think there ever
have been - so remove it from the `export` statement.