Refactor to use `withOidcSecure` HOC
The `@axa-fr/react-oidc` library provides two approaches for requiring a
logged-in user: the `<OidcSecure>` component, and the `withOidcSecure`
higher-order component (HOC). `<OidcSecure>` has the advantages of being
a little simpler conceptually, and it lets us avoid naming the insecure
inner component.

However, the component approach also requires that every return path
include an `<OidcSecure>`, which gets to be redundant and causes extra
indentation. As we start building more complex return values - as we'll
be doing shortly to support the proposal list sidebar - those drawbacks
become more severe.

The HOC allows the inner component to not have to worry about security,
simplifying the React nodes they return.

Introduce the suffix `Loader` for those to-be-secured components that
make API calls, parse the results, and pass data into display
components, and wrap those loaders with `withOidcSecure`.

While this does affect the React component tree, by moving the location
of the `OidcSecure` component up a level, this should have no effect on
application functionality.

Issue #62 Add sidebar to Proposal Detail page