Add proposal list sidebar to proposal detail page
Use the new `<ProposalListGridPanel>` component in a sidebar on the
proposal detail page.

Since both components need the base (canonical) field list, fetch it
from the parent component and pass it to both, so that we don't
needlessly repeat API calls.

For now, hardcode the page number and page size for the sidebar list, as
we don't have proper pagination (or server side search) yet, and it's
worth spending a bit more time considering how best to wire that up.

Issue #62 Add sidebar to Proposal Detail page