Rename API interface type definitions
In commit 056b45c5e9c2ee3acc7cd2a17ebf566681530961, we introduced a name
for the display-centric models we are using. The name introduced,
`DataViewerProposal`, is prefixed to distinguish itself from the type
definitions for how the API represents data.

Rename the API-centric interfaces to include an `Api` prefix.

These new names make it clearer that we should not be using these
interfaces so widely, but cleaning that up will be part of future work.

Issue #108 Rename shared TypeScript interfaces for clarity