Use server search API
Rather than always loading the full list of proposals from the API and
doing basic string matching locally, instead use the recently introduced
search API. The server search API includes such fanciness as word
stemming (so that a query for "summaries" would match a proposal
containing "summary") and operators (like "a OR b" which would match
proposals containing either).

Note that we have not yet implemented pagination, so this is somewhat of
a tradeoff: the initial page load will get the full list of proposals,
and searching will cause another server round trip, and clearing the
search will cause a third round trip. However, this is a necessary step
on the way to pagination.

Issue #15 Implement pagination on proposal lists
Issue #83 Enable search