Extract constants for proposals API default values
We are showing two lists of proposals, one in `<ProposalListLoader>` and
one in `<ProposalListGridLoader>`. Each of these needs to call the
`/proposals` API, and so needs to specify the page number and how many
proposals to show per page - and, shortly, the search query.

Extract constants for the default values we're passing to the API -
which, it may be worth noting, differ from the default values the API
uses when none are passed - and use them in both components that need
the `useProposals` hook.

Issue #15 Implement pagination on proposal lists
Issue #83 Enable search