Add `mapProposals()` tests
This commit adds tests of the `mapProposals()` function.

It also adds a bare minimum mock data structure for canonical fields and
API proposals; I recognize we may not actually want to define mock data
in the tests like this, but I think it would be trivial to replace this
down the road, and the more important thing is testing these functions.