Use applicant name cascade on proposal table
Prior to this commit, the proposal table’s Organization Name column was
rendering precisely what was in the `organization_name` field. However,
in that view (as in the sidebar and atop the proposal detail page), we
actually want to cascade through a list of possible fields that might
contain the applicant name.

(Side note: we visibly say “Organization Name” a lot, but the proper
terminology for this entity is “applicant”, so in code I will refer to
it that way.)

Now, we dynamically generate the cell contents from the applicant name
candidate fields, which ensures we always have a value in that cell —
even if it’s the fallback value.

Note that `getProposalCellContents()` is implemented with a switch
statement because I predict we’ll do something similar for other columns
very soon. We may not stick with this syntax but it’s a good bet.

Issue #109 Display applicant name from candidate fields