Show loading screen when fetching new info
Previously, navigating to another instance of the same page (in the
sense of react-router) would show the old data until the API request for
new data succeeded. This occured both on the proposals list page with
search, and on the proposal detail page when clicking on another
proposal in the sidebar list or clicking the dev-mode-only prev/next
buttons.

This behavior is incorrect: if we navigate from `/proposals/27` to
`/proposals/33`, we shouldn't show the details for proposal 27 after the
URL has changed. Instead, we should do the same thing we'd do if we
navigated from `/proposals`, or if we refreshed the page: show the
loading screen until the proposal 33 data has loaded.

Invalidate the old data as the first step in retrieving new data, so
that the existing loading screens are shown appropriately.

Issue #83 Enable search