Show proposal list panel while loading
This commit adds a `loading` boolean to the `ProposalListTablePanel`
component so that we can go ahead and render it even while waiting for
its data from the API. This means the interface appears more stable for
the user, with data filling in to the panel as it arrives, rather than
the entire panel waiting to render until data is available.

While the data is loading, we show either "Loading…" or "Searching…"
depending on the situation.

Issue #83 Enable search