Fix applicant name on proposal detail header
Prior to this commit, the proposal detail page would sometimes display
a blank applicant name atop the main panel.

That’s because this page uses the API’s proposal data directly as it
came from the API, without running it through our the `mapProposal()`
function that — among other things — strips out blank values. Thus, when
when the API populated an applicant name field with a blank string, that
blank string would get “displayed” on the proposal detail page. Which is
wrong.

Long-term, the API should probably not return blank strings, which
aren’t meaningful values. For now, we should just disregard blank values
when selecting an applicant name.

This commit does so, but scopes the fix to only the `getApplicant()`
function because we want to be very conscious of when we’re doing this.

Issue #109 Display applicant name from candidate fields