Don’t let panel tag icons shrink
Previously, if the panel was narrow enough to e.g. cause the panel tag
to wrap, the icon would shrink at the same rate that the text portion
shrank. But the icon only occupies a few tens of pixels of width;
shrinking it helps very little, at the cost of icon legibility and
usability.

Don’t let it shrink. Force the text to wrap.