Move overflow scrolling from panel to panel body
Prior to this commit, when the panel body was too large for its parent
panel, the body grew as per normal flow, increasing the internal size of
the parent and causing the parent’s overflow:auto to trigger.
Consequently, when the panel body was too wide, it caused the panel
header to grow along with it. This was not ideal!

We actually want the panel body to independently scroll, which it does
now. We then had to make a slight tweak to the sticky-position math on
tables, so that they were only negative-offset when they were sticking
to the top of a panel rather than to the top of a panel body.

Issue #86 Panel actions don't stay fixed in top corners when panel
          overflows