Extract <ClosablePanel> component
There are going to be several different versions of the right side panel
we'll want to use, primarily to handle different error cases with
showing data platform provider information. Rather than duplicate the
functionality in each, create a single component that knows how to close
itself, so that we'll be able to use that as the basis for each
component. Start with the existing <DataPlatformProviderPanel>
component.

Issue #23 Create applicant detail page