Relax dangling comma rule in multiline functions
Loosen the comma-dangle rule (which we have via the TypeScript
version[1], which extends the base version[2]) to not require trailing
commas on multiline functions. In particular, that rule is onerous when
a single-argument function is spilled across multiple lines, and
especially when the argument is itself a multiline expression.

(What I really want, but which ESLint does not support, is a mode which
requires dangling commas on functions that have two or more arguments,
and allows but does not require them on functions that take only a
single argument.)

[1] https://typescript-eslint.io/rules/comma-dangle/
[2] https://eslint.org/docs/latest/rules/comma-dangle