Close dropdown when menu links are clicked
Currently, clicking on a dropdown menu item takes no action on the
dropdown menu itself. When the dropdown menu item is a link that causes
the page to re-render in a way that unmounts the dropdown menu, that’s
no big deal. But when the dropdown menu remains on screen, the menu
remains open. We don’t want that!

This change closes the ancestor dropdown when a dropdown menu link is
clicked. Note that it uses the native DOM traversal because passing by
ref would be super noisy here. Also note that this is not currently
configurable; a future change could add this, but @jasonaowen suggests
that this component would still be the correct configuration point so
that items that do close the menu could be intermixed with those that
didn’t.

Issue #190 Close dropdown when item is used