Show Charity Navigator data
Add parsing for Charity Navigator data.

The responses Charity Navigator has for the organizations whose data we
are working with, and the fields we want to show, contain relatively
more non-string types - including sometimes null. Fix the data type
we're asserting JSONPath is returning to reflect that, filter out null
values, and cast the rest to strings.

Issue #23 Create applicant detail page