Show Candid data
Fetch the cached data platform provider information from the PDC, and
display the Candid data.

To do so, use JSON paths via the jsonpath-plus[1] library. This allows
us to specify a series of paths to pull data out of the large and
complex Candid response, with relatively graceful error handling: if a
path is not found, no data is displayed.

jsonpath-plus also returns JSON pointers as part of its "all" result
format, which lend themselves fairly well to automatically-generated
labels for our UI. This is a slightly brittle approach, as it would be
difficult to add a hand-crafted label into the mix, but for now it seems
to meet our needs.

Select several pieces of data from the Candid response, based on
conversations with @jim-mcgowan.

[1] https://jsonpath-plus.github.io/JSONPath/docs/ts/index.html

Issue #23 Create applicant detail page