Add `padded` attribute to PanelBody
This commit adds a default padding to PanelBody components, configurable
via the `padded` prop. (This prop replaces the old `fill` prop that was
not actually used, but was meant to imply the inverse of `padded`.) It
also required adding the attribute to `ClosablePanel`, since they need
to be able to pass it on through to the `PanelBody` insde.

It also updates all uses of `PanelBody` (and `ClosablePanel`) to
indicate whether they should be padded or not.

Issue #195 Panel bodies have no internal padding