Remove global transition-duration property
In an excellent find by @jasonaowen, my attempt at setting a default
global transition duration — so that individual elements could define
which properties they wanted to transition without having to re-declare
“just use the global transition speed” — implicitly enables transitions
of all properties for all elements. This is not what we want, and in
particular was causing the change of a `PanelBody` between `padded`
configurations to be improperly transitioned rather than instant.

Tagged with the issue during which this commit was developed, but
actually a larger fix.

Issue #195 Panel bodies have no internal padding